requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
duckduckgo_search_api==0.1.5
googlesearch-python==1.3.0
gunicorn==23.0.0
idna==3.10
Levenshtein==0.27.1
packaging==25.0
RapidFuzz==3.14.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
